backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.5
insightface==0.7.3
onnxruntime==1.19.2
numpy==1.24.3
Pillow==10.0.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
aiofiles==23.2.1
requests==2.31.0
